✨ Add Feature: Random Cafe Selections (#4)
This commit adds the ability for the user to ask the bot to randomly
recommend a cafe for them to visit.

The functionality is triggered by the user if the bot is mentioned in a
message with the word "coffee" in it. The bot will then ask if there is
a preference on the location of the cafe, and will then randomly select
a cafe from the list of cafes in the cafe list.