🐛 Ignore Old AppMentions Events
This commit adds a check to the AppMention event handler, to check how
recent the message was created. If the message is greater than 2 seconds
old, it will be ignored. This will prevent the bot responding to edited,
older messages.

Previously, the bot would respond to any incoming AppMention event,
regardless of how old the message was.

Fixes #16.